requirements.txt
area==1.1.1
beautifulsoup4==4.9.3
certifi==2021.5.30
charset-normalizer==2.0.4
geojson==2.5.0
idna==3.2
numpy==1.21.1
Pillow==8.3.1
progressbar2==3.53.1
pycoral==2.0.0
python-decouple==3.4
python-utils==2.5.6
requests==2.26.0
six==1.16.0
soupsieve==2.2.1
tflite-runtime==2.5.0.post1
urllib3==1.26.6
